refactor geopandas dtypes to avoid top-level import
