+++ pandera/typing/geopandas.py
@@ -39,7 +39,7 @@ except ImportError:  # pragma: no cover
 
 if GEOPANDAS_INSTALLED:
     # pylint: disable=import-outside-toplevel,ungrouped-imports
-    from pandera.engines.pandas_engine import Geometry
+    from pandera.engines.geopandas_engine import Geometry
 
     # pylint:disable=invalid-name
     if TYPE_CHECKING:
