+++ tests/strategies/test_strategies.py
@@ -16,7 +16,7 @@ from pandera import strategies
 from pandera.api.checks import Check
 from pandera.api.extensions import register_check_statistics
 from pandera.dtypes import is_category, is_complex, is_float
-from pandera.engines import pandas_engine
+from pandera.engines import pandas_engine, geopandas_engine
 
 try:
     import hypothesis
@@ -88,8 +88,8 @@ for data_type in pandas_engine.Engine.get_registered_dtypes():
         or is_category(data_type)
         or data_type in UNSUPPORTED_DTYPE_CLS
         or (
-            pandas_engine.GEOPANDAS_INSTALLED
-            and data_type == pandas_engine.Geometry
+            geopandas_engine.GEOPANDAS_INSTALLED
+            and data_type == geopandas_engine.Geometry
         )
     ):
         continue
